tried adding physics
tried adding physics to the game and to make a program that moves the enemy paddle, altough i think ill just put a wall so the player can play against the wall. right now the code's pretty buggy, meaning that the enemy paddle is not moving, as its following the ball. im thinking of a way to start the game. tried asking chatgpt for help, but it gave me garbage (garbage as in a code that doesnt work, the whole screen went black) and so, for now at least, no more chatgpt! yay.....